ai_service/requirements.txt
fastapi
uvicorn
pytesseract
pillow
openai
python-multipart
python-dotenv
